fix type comparaison

+++ project/scripts/str.c
@@ -97,7 +97,7 @@ int str_to_int(char *str)
         return -1;
     }
     int result = 0;
-    for (int i = 0; i < strlen(str); i++)
+    for (size_t i = 0; i < strlen(str); i++)
     {
         result += (str[i]-'0')*power(10, strlen(str)-i-1);
     }
